feat: organize code into packages
