+++ go.mod
@@ -2,15 +2,15 @@ module github.com/appwrite/monitoring
 
 go 1.19
 
-require github.com/shirou/gopsutil/v3 v3.24.1
+require github.com/shirou/gopsutil/v3 v3.23.7
 
 require (
 	github.com/go-ole/go-ole v1.2.6 // indirect
 	github.com/lufia/plan9stats v0.0.0-20211012122336-39d0f177ccd0 // indirect
 	github.com/power-devops/perfstat v0.0.0-20210106213030-5aafc221ea8c // indirect
 	github.com/shoenig/go-m1cpu v0.1.6 // indirect
-	github.com/tklauser/go-sysconf v0.3.12 // indirect
-	github.com/tklauser/numcpus v0.6.1 // indirect
+	github.com/tklauser/go-sysconf v0.3.11 // indirect
+	github.com/tklauser/numcpus v0.6.0 // indirect
 	github.com/yusufpapurcu/wmi v1.2.3 // indirect
-	golang.org/x/sys v0.16.0 // indirect
+	golang.org/x/sys v0.10.0 // indirect
 )
