+++ main.go
@@ -1,320 +1,16 @@
 package main
 
 import (
-	"encoding/json"
 	"flag"
 	"fmt"
-	"net/http"
 	"os"
-	"path/filepath"
-	"strings"
-	"time"
 
-	"github.com/shirou/gopsutil/v3/cpu"
-	"github.com/shirou/gopsutil/v3/disk"
-	"github.com/shirou/gopsutil/v3/mem"
+	"github.com/appwrite/monitoring/pkg"
+	"github.com/appwrite/monitoring/pkg/monitor"
 )
 
-type Metric struct {
-	Title     string  `json:"title"`
-	Cause     string  `json:"cause"`
-	AlertID   string  `json:"alert_id"`
-	Timestamp int64   `json:"timestamp"`
-	Status    string  `json:"status"`
-	Value     float64 `json:"value"`
-	Limit     float64 `json:"limit"`
-}
-
-type SystemMonitor struct {
-	httpClient     *http.Client
-	betterStackURL string
-	hostname       string
-	cpuLimit       float64
-	memoryLimit    float64
-	diskLimit      float64
-	interval       int
-	log            *Logger
-	
-	// EMA tracking
-	cpuEMA         float64
-	memoryEMA      float64
-	diskEMAs       map[string]float64 // Map to track EMAs for all disks (root and mounted)
-	alpha          float64 // EMA smoothing factor
-}
-
-func NewSystemMonitor(betterStackURL string, interval int, cpuLimit, memoryLimit, diskLimit float64) (*SystemMonitor, error) {
-	hostname, err := os.Hostname()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get hostname: %v", err)
-	}
-
-	// Calculate alpha based on interval to get roughly 5 minutes of smoothing
-	// EMA formula: alpha = 2/(N+1) where N is the number of periods
-	// For 5 minutes of smoothing with our interval: N = 300/interval
-	N := float64(300) / float64(interval)
-	alpha := 2.0 / (N + 1.0)
-
-	return &SystemMonitor{
-		httpClient: &http.Client{
-			Timeout: 5 * time.Second,
-		},
-		betterStackURL: betterStackURL,
-		hostname:       hostname,
-		cpuLimit:       cpuLimit,
-		memoryLimit:    memoryLimit,
-		diskLimit:      diskLimit,
-		interval:       interval,
-		log:            New(),
-		diskEMAs:       make(map[string]float64), // Initialize the map for all disk EMAs
-		alpha:          alpha,
-	}, nil
-}
-
-func (s *SystemMonitor) calculateEMA(currentValue, previousEMA float64) float64 {
-	return s.alpha*currentValue + (1-s.alpha)*previousEMA
-}
-
-func (s *SystemMonitor) checkCPU() error {
-	duration := float64(s.interval) / 10
-	if duration < 5 {
-		duration = 5
-	}
-	if duration > 60 {
-		duration = 60
-	}
-
-	cpuPercent, err := cpu.Percent(time.Duration(duration)*time.Second, false)
-	if err != nil {
-		return fmt.Errorf("failed to get CPU usage: %v", err)
-	}
-
-	if len(cpuPercent) == 0 {
-		return nil
-	}
-
-	// Calculate EMA for CPU usage
-	instantValue := cpuPercent[0]
-	s.cpuEMA = s.calculateEMA(instantValue, s.cpuEMA)
-	
-	status := s.getStatus(s.cpuEMA, s.cpuLimit)
-	if status == "fail" {
-		s.log.Warn("CPU usage EMA %.2f%% exceeds limit of %.2f%% (instant: %.2f%%)", s.cpuEMA, s.cpuLimit, instantValue)
-	} else {
-		s.log.Log("CPU usage EMA: %.2f%% (limit: %.2f%%, instant: %.2f%%)", s.cpuEMA, s.cpuLimit, instantValue)
-	}
-	
-	metric := Metric{
-		Title:     fmt.Sprintf("CPU Usage - %s", s.hostname),
-		Cause:     "CPU monitoring check",
-		AlertID:   fmt.Sprintf("cpu-%s", s.hostname),
-		Timestamp: time.Now().Unix(),
-		Status:    status,
-		Value:     s.cpuEMA,
-		Limit:     s.cpuLimit,
-	}
-
-	return s.sendMetric(metric)
-}
-
-func (s *SystemMonitor) checkMemory() error {
-	vmStat, err := mem.VirtualMemory()
-	if err != nil {
-		return fmt.Errorf("failed to get memory stats: %v", err)
-	}
-
-	instantValue := vmStat.UsedPercent
-	s.memoryEMA = s.calculateEMA(instantValue, s.memoryEMA)
-	
-	status := s.getStatus(s.memoryEMA, s.memoryLimit)
-	if status == "fail" {
-		s.log.Warn("Memory usage EMA %.2f%% exceeds limit of %.2f%% (instant: %.2f%%)", s.memoryEMA, s.memoryLimit, instantValue)
-	} else {
-		s.log.Log("Memory usage EMA: %.2f%% (limit: %.2f%%, instant: %.2f%%), Available: %d MB, Total: %d MB",
-			s.memoryEMA,
-			s.memoryLimit,
-			instantValue,
-			vmStat.Available/(1024*1024),
-			vmStat.Total/(1024*1024))
-	}
-
-	metric := Metric{
-		Title:     fmt.Sprintf("Memory Usage - %s", s.hostname),
-		Cause:     "Memory monitoring check",
-		AlertID:   fmt.Sprintf("memory-%s", s.hostname),
-		Timestamp: time.Now().Unix(),
-		Status:    status,
-		Value:     s.memoryEMA,
-		Limit:     s.memoryLimit,
-	}
-
-	return s.sendMetric(metric)
-}
-
-func (s *SystemMonitor) checkDisk() error {
-	// Check root partition
-	rootPath := "/"
-	usage, err := disk.Usage(rootPath)
-	if err != nil {
-		return fmt.Errorf("failed to get disk usage: %v", err)
-	}
-
-	instantValue := usage.UsedPercent
-	
-	// Calculate or update EMA for root disk
-	if _, exists := s.diskEMAs[rootPath]; !exists {
-		// Initialize EMA with current value if this is first check
-		s.diskEMAs[rootPath] = instantValue
-	}
-	
-	// Update EMA for root disk
-	s.diskEMAs[rootPath] = s.calculateEMA(instantValue, s.diskEMAs[rootPath])
-	
-	rootEMA := s.diskEMAs[rootPath]
-	status := s.getStatus(rootEMA, s.diskLimit)
-	
-	if status == "fail" {
-		s.log.Warn("Root disk usage EMA %.2f%% exceeds limit of %.2f%% (instant: %.2f%%)", rootEMA, s.diskLimit, instantValue)
-	} else {
-		s.log.Log("Root disk usage EMA: %.2f%% (limit: %.2f%%, instant: %.2f%%), Free: %d MB, Total: %d MB",
-			rootEMA,
-			s.diskLimit,
-			instantValue,
-			usage.Free/(1024*1024),
-			usage.Total/(1024*1024))
-	}
-
-	if err := s.sendMetric(Metric{
-		Title:     fmt.Sprintf("Root Disk Usage - %s", s.hostname),
-		Cause:     "Disk monitoring check",
-		AlertID:   fmt.Sprintf("disk-root-%s", s.hostname),
-		Timestamp: time.Now().Unix(),
-		Status:    status,
-		Value:     rootEMA,
-		Limit:     s.diskLimit,
-	}); err != nil {
-		return err
-	}
-
-	// Check mounted directories
-	mounts, err := filepath.Glob("/mnt/*")
-	if err != nil {
-		return fmt.Errorf("failed to list mounted directories: %v", err)
-	}
-
-	for _, mount := range mounts {
-		usage, err := disk.Usage(mount)
-		if err != nil {
-			s.log.Error("Failed to get disk usage for %s: %v", mount, err)
-			continue
-		}
-
-		instantValue := usage.UsedPercent
-		
-		// Calculate or update EMA for this mount
-		if _, exists := s.diskEMAs[mount]; !exists {
-			// Initialize EMA with current value if this is first check
-			s.diskEMAs[mount] = instantValue
-		}
-		
-		// Update EMA for this mount
-		s.diskEMAs[mount] = s.calculateEMA(instantValue, s.diskEMAs[mount])
-		
-		mountEMA := s.diskEMAs[mount]
-		status := s.getStatus(mountEMA, s.diskLimit)
-		
-		if status == "fail" {
-			s.log.Warn("Disk usage for %s EMA %.2f%% exceeds limit of %.2f%% (instant: %.2f%%)", mount, mountEMA, s.diskLimit, instantValue)
-		} else {
-			s.log.Log("Disk usage for %s EMA: %.2f%% (limit: %.2f%%, instant: %.2f%%), Free: %d MB, Total: %d MB",
-				mount,
-				mountEMA,
-				s.diskLimit,
-				instantValue,
-				usage.Free/(1024*1024),
-				usage.Total/(1024*1024))
-		}
-
-		if err := s.sendMetric(Metric{
-			Title:     fmt.Sprintf("Disk Usage %s - %s", mount, s.hostname),
-			Cause:     "Disk monitoring check",
-			AlertID:   fmt.Sprintf("disk-%s-%s", filepath.Base(mount), s.hostname),
-			Timestamp: time.Now().Unix(),
-			Status:    status,
-			Value:     mountEMA,
-			Limit:     s.diskLimit,
-		}); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (s *SystemMonitor) getStatus(value, limit float64) string {
-	if value > limit {
-		return "fail"
-	}
-	return "pass"
-}
-
-func (s *SystemMonitor) sendMetric(metric Metric) error {
-	body, err := json.Marshal(metric)
-	if err != nil {
-		return fmt.Errorf("failed to marshal metric: %v", err)
-	}
-
-	req, err := http.NewRequest(http.MethodPost, s.betterStackURL, strings.NewReader(string(body)))
-	if err != nil {
-		return fmt.Errorf("failed to create request: %v", err)
-	}
-
-	req.Header.Set("Content-Type", "application/json; charset=utf-8")
-	req.Header.Set("Accept", "application/json")
-	req.Header.Set("User-Agent", "Appwrite Resource Monitoring")
-
-	resp, err := s.httpClient.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to send request: %v", err)
-	}
-	defer resp.Body.Close()
-
-	s.log.Log("Response Status: %s", resp.Status)
-	if resp.StatusCode >= 400 {
-		return fmt.Errorf("request failed with status: %d", resp.StatusCode)
-	}
-
-	return nil
-}
-
-func (s *SystemMonitor) Start() {
-	ticker := time.NewTicker(time.Duration(s.interval) * time.Second)
-	defer ticker.Stop()
-
-	// Initial check
-	s.runChecks()
-
-	// Periodic checks
-	for range ticker.C {
-		s.runChecks()
-	}
-}
-
-func (s *SystemMonitor) runChecks() {
-	if err := s.checkCPU(); err != nil {
-		s.log.Error("Error checking CPU: %v", err)
-	}
-
-	if err := s.checkMemory(); err != nil {
-		s.log.Error("Error checking memory: %v", err)
-	}
-
-	if err := s.checkDisk(); err != nil {
-		s.log.Error("Error checking disk: %v", err)
-	}
-}
-
 func main() {
-	log := New()
+	log := pkg.NewLogger()
 
 	// Command line flags
 	betterStackURL := flag.String("url", "", "BetterStack webhook URL (required)")
@@ -351,7 +47,7 @@ func main() {
 		log.Fatal("Disk limit must be between 0 and 100")
 	}
 
-	monitor, err := NewSystemMonitor(*betterStackURL, *interval, *cpuLimit, *memoryLimit, *diskLimit)
+	monitor, err := monitor.NewSystemMonitor(*betterStackURL, *interval, *cpuLimit, *memoryLimit, *diskLimit)
 	if err != nil {
 		log.Fatal("Failed to create system monitor: %v", err)
 	}
